tests/wpt/web-platform-tests/css/css-fonts/parsing/font-weight-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Module Level 3: parsing font-weight with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-fonts-3/#font-weight-prop">
<meta name="assert" content="font-weight supports the full grammar 'normal | bold | bolder | lighter | 100 | 200 | 300 | 400 | 500 | 600 | 700 | 800 | 900'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('font-weight', 'normal');
test_valid_value('font-weight', 'bold');
test_valid_value('font-weight', 'bolder');
test_valid_value('font-weight', 'lighter');
test_valid_value('font-weight', '100');
test_valid_value('font-weight', '200');
test_valid_value('font-weight', '300');
test_valid_value('font-weight', '400');
test_valid_value('font-weight', '500');
test_valid_value('font-weight', '600');
test_valid_value('font-weight', '700');
test_valid_value('font-weight', '800');
test_valid_value('font-weight', '900');
</script>
</body>
</html>
